+++ include/options.hpp
@@ -53,6 +53,7 @@ namespace ghost
 		int tabu_time_selected; //!< Number of local moves a selected variable is marked tabu.
 		int percent_chance_force_trying_on_plateau; //!< Percentage of chance to force trying another variable rather than doing a move on a plateau.
 		int max_stay_on_plateau; //!< Maximal number of local moves in a row the solver does on a plateau.
+		int max_moves_in_opt_space; //!< Maximal number of local moves in an optimization space.
 		int reset_threshold; //!< Number of variables marked as tabu required to trigger a reset.
 		int restart_threshold; //!< Trigger a restart every 'restart_threshold' reset. Set to 0 to never trigger restarts.
 		int number_variables_to_reset; //!< Number of variables to randomly change the value at each reset.